# Assignment count refactor