--- a/juntagrico/dao/assignmentdao.py
+++ b/juntagrico/dao/assignmentdao.py
@@ -1,10 +1,6 @@
-from datetime import datetime, time
-
 from django.utils import timezone
-from django.utils.timezone import get_default_timezone as gdtz
 
 import juntagrico
-from juntagrico.util.temporal import start_of_business_year
 
 
 class AssignmentDao:
@@ -20,12 +16,6 @@ class AssignmentDao:
     @staticmethod
     def assignments_for_member(member):
         return juntagrico.entity.jobs.Assignment.objects.filter(member=member)
-
-    @staticmethod
-    def assignments_for_member_current_business_year(member):
-        start = datetime.combine(start_of_business_year(), time.min, tzinfo=gdtz())
-        return juntagrico.entity.jobs.Assignment.objects.filter(member=member).\
-            filter(job__time__gte=start, job__time__lt=timezone.now())
 
     @staticmethod
     def upcomming_assignments_for_member(member):