--- a/juntagrico/entity/subs.py
+++ b/juntagrico/entity/subs.py
@@ -1,5 +1,3 @@
-from datetime import timedelta
-
 from django.contrib import admin
 from django.db import models
 from django.db.models import Q, QuerySet, F, Sum
@@ -15,8 +13,9 @@ from juntagrico.entity.member import q_left_subscription, q_joined_subscription
 from juntagrico.entity.subtypes import SubscriptionType
 from juntagrico.lifecycle.sub import check_sub_consistency
 from juntagrico.lifecycle.subpart import check_sub_part_consistency
+from juntagrico.queryset.subscription import SubscriptionQuerySet
 from juntagrico.util.models import q_activated, q_cancelled, q_deactivated, q_deactivation_planned, q_isactive
-from juntagrico.util.temporal import start_of_next_business_year, start_of_business_year, end_of_business_year
+from juntagrico.util.temporal import start_of_next_business_year
 
 
 class Subscription(Billable, SimpleStateModel):
@@ -42,6 +41,8 @@ class Subscription(Billable, SimpleStateModel):
     notes = models.TextField(
         _('Notizen'), max_length=1000, blank=True,
         help_text=_('Notizen für Administration. Nicht sichtbar für {}'.format(Config.vocabulary('member'))))
+
+    objects = SubscriptionQuerySet.as_manager()
 
     def __str__(self):
         return _('Abo ({1}) {0}').format(self.size, self.id)
@@ -80,11 +81,6 @@ class Subscription(Billable, SimpleStateModel):
     @property
     def active_and_future_parts(self):
         return self.parts.filter(~q_deactivated())
-
-    @property
-    def parts_in_business_year(self):
-        return self.parts.filter(Q(activation_date__isnull=False, activation_date__lte=end_of_business_year()) &
-                                 ~Q(deactivation_date__isnull=False, deactivation_date__lt=start_of_business_year()))
 
     @property
     def part_change_date(self):
@@ -128,16 +124,12 @@ class Subscription(Billable, SimpleStateModel):
     def subscription_amount_future(self, size):
         return self.calc_subscription_amount(self.future_parts, size)
 
-    @property
-    def required_assignments(self):
-        return self.get_required_assignments()
-
-    @property
-    def required_core_assignments(self):
-        return self.get_required_assignments(True)
-
-    def get_required_assignments(self, core=False):
-        return round(sum([part.get_required_assignments(core) for part in self.parts_in_business_year.select_related('type')]))
+    def get_with_assignments(self, *args, **kwargs):
+        """
+        calculate assignments of this sub and write result into properties.
+        If applying this on multiple elements use `annotate_assignments()` on queryset instead.
+        """
+        return Subscription.objects.annotate_assignments(*args, **kwargs).get(pk=self)
 
     @property
     def price(self):
@@ -295,23 +287,6 @@ class SubscriptionPart(JuntagricoBaseModel, SimpleStateModel):
     @property
     def is_extra(self):
         return self.type.size.product.is_extra
-
-    def get_required_assignments(self, core=False):
-        if not self.activation_date:
-            return 0
-        nominal_required = self.type.required_core_assignments if core else self.type.required_assignments
-        # since when part is active in current business year
-        start_business = start_of_business_year()
-        start = max(self.activation_date, start_business)
-        # when (if at all) part will be inactive in current business year
-        end_business = end = end_of_business_year()
-        if self.deactivation_date:
-            end = min(self.deactivation_date, end_business)
-        elif self.type.trial_days:
-            end = min(self.activation_date + timedelta(self.type.trial_days - 1), end_business)
-        # percentage of business year (or trial period), where part is active. Do not round here.
-        period = timedelta(self.type.trial_days) or (end_business - start_business + timedelta(1))
-        return nominal_required * (end - start + timedelta(1)) / period
 
     def clean(self):
         check_sub_part_consistency(self)