--- a/testsettings.py
+++ b/testsettings.py
@@ -36,6 +36,18 @@ DATABASES = {
     }
 }
 
+if os.environ.get('GITHUB_WORKFLOW'):
+    DATABASES = {
+        'default': {
+           'ENGINE': 'django.db.backends.postgresql',
+           'NAME': 'testdb',
+           'USER': 'postgres',
+           'PASSWORD': 'postgres',
+           'HOST': '127.0.0.1',
+           'PORT': '5432',
+        }
+    }
+
 ROOT_URLCONF = 'testurls'
 
 AUTHENTICATION_BACKENDS = (