… .github/workflows/juntagrico-ci.yml
… juntagrico/templates/management_lists/assignments.html
… juntagrico/templatetags/juntagrico/widgets.py
… juntagrico/util/subs.py
… juntagrico/util/temporal.py
… juntagrico/views_admin.py
… requirements-local.txt
… test/test_assignments.py
… test/test_cs.py
… test/util/test.py