Initial motivation was to calculate assignments for the "last year", i.e. have a simple method to say: attach assignment data between this start and end date to this set of assignments.
Before some calculations (required assignments) were properties on the Subscription (which is slow for large data sets) and others were collected into a **list** with `subscriptions_with_assignments`.
Queryset annotations seem to be the most powerful approach. That meant moving calculations to the database. A good opportunity to learn more about databases ;-)
There is an alternative approach if we reach the limits of annotations one day (which I doubt after this one): Use a proxy model with the extra properties and a custom base manager that applies the required annotations by default there.
I also started using custom querysets instead of dao as this is the path forward #527

# Features
- Enable assignment counting for arbitrary periods
- Count assignments of future jobs or don't, using a simple `count_jobs_until` argument
- Refactor: Move all assignment counting logic to one place, annotate all assignments data with a few methods
- Speed improvements: Faster assignment overview and export

# Bugfixes

- Don't count assignments of members that have left a subscription
- Fix double counting of assignments of members that switch subscription mid year. Assignments are now counted according to when the job was done and join/leave dates of the subscription membership
- Fix sorting of assignment counts with decimals in management list.

# Todo

- [x] Run automated tests with postgres and other supported databases. Only test postgres for now, MySQL has a bug #423.
- [x] Compare performance to previous implementation
- [x] Try to make `annotate_assignments_progress` independent of the other annotations (i.e. annotate internally)
- [x] Modernise `MemberDao.annotate_members_with_assignemnt_count` and check if it can be reused. #599