@motu55 2023-02-06
just a little suggestion

what you are actually implementing is that assignments belong to the subscription.

so you could instad of the subquery over subscription membership add a nullable foreign key to the assignemnts which points to a subscription.

This would make some of the queries more lightweight and it would allow organisations in certain cases to move assignemnts to a new subscription because of certain unforseen circumstances (something which woul no mean to create fake subscriptionmemberships which destroys the integrity of the subscription membership concept and the whole contept of seeing who did exactly what when)

Also it would allow in the future to implement an easy switch so that organisations can choose if the assignments belong to a member or a subscription