@pascalfree 2023-02-06
The idea for assignments on subscriptions is also captured here: https://github.com/juntagrico/juntagrico/issues/521#issuecomment-1059850238
I agree that this will be the way to go.

Here I did not yet change the counting logic really. I only fixed double counting. For members that switch sub during the year their assignments counted to the old and new sub for this year.