@pascalfree 2023-02-14
I did some performance comparisons. On a local machine using postgres with a real world data set of around 300 active subscriptions.

I loaded the assignments overview page, but removed the comembers (and their emails) in previous and new version to reduce the noise. (scenario 1). After seeing the queries I found that it could be optimized by pre-selecting the primary member (select_related) and did that for both versions as well (scenario 2). Finally I loaded the contact page to just the effect of the assignment widget.

# Results

 scenario | before | after 
--------|-------------|------------
No comembers | CPU: 33s <br> SQL: 2353 in 2063ms | CPU: 8s <br> SQL: 552 in 486ms 
select related | CPU: 26s <br>  SQL: 2062 in 1237ms | CPU: 1.6s <br> SQL: 38 in 338ms 
contact page | CPU: 460ms <br> SQL: 141 in 67ms | CPU: 337ms <br> SQL: 32 in 83ms 

# Interpretation
For the assignments management table speed increase is significant and number of SQL queries reduces as well. As expected.
For the widget alone also the number of queries was reduced but the time on executing them increased. Still the request is faster overall, as the calculations have been moved to the db.

# Conclusion
The new version is faster