@pascalfree 2023-02-20
> I would rather add a froeign key on subscription on the assignement than the whole subscriptionmembership magic. as stated before this gives admins a bit more flexibility and reduces the query by quite a lot since no join is nescessary on subscription membership.

This will raise a bunch of other questions (like, how do I display this?) that need to be adressed to do a proper implementation and excedes the scope of this refactor. The subscriptionmembership magic was there before, I just moved it to another place a fixed a bug in it.
I'll merge this and will implement the foreign key later in https://github.com/juntagrico/juntagrico/issues/521#issuecomment-1059850238